+++ examples/utfgrid.html
@@ -1,11 +1,11 @@
 ---
 layout: example.html
-title: Tiled UTFGrid
-shortdesc: This example shows how to read data from a TileUTFGrid layer.
+title: UTFGrid
+shortdesc: This example shows how to read data from a UTFGrid source.
 docs: >
   <p>Point to a country to see its name and flag.</p>
   Tiles made with [TileMill](http://tilemill.com). Hosting on MapBox.com or with open-source [TileServer](https://github.com/klokantech/tileserver-php/).
-tags: "utfgrid, tileutfgrid, tilejson"
+tags: "utfgrid, tilejson"
 cloak:
   - key: pk.eyJ1IjoiYWhvY2V2YXIiLCJhIjoiRk1kMWZaSSJ9.E5BkluenyWQMsBLsuByrmg
     value: Your Mapbox access token from http://mapbox.com/ here
