+++ src/ol/source.js
@@ -20,9 +20,9 @@ export {default as TileArcGISRest} from './source/TileArcGISRest.js';
 export {default as TileDebug} from './source/TileDebug.js';
 export {default as TileImage} from './source/TileImage.js';
 export {default as TileJSON} from './source/TileJSON.js';
-export {default as TileUTFGrid} from './source/TileUTFGrid.js';
 export {default as TileWMS} from './source/TileWMS.js';
 export {default as UrlTile} from './source/UrlTile.js';
+export {default as UTFGrid} from './source/UTFGrid.js';
 export {default as Vector} from './source/Vector.js';
 export {default as WMTS} from './source/WMTS.js';
 export {default as XYZ} from './source/XYZ.js';
